Seperated out buffer and plugin tests
